va: split unit tests from va_test.go
